+++ shardingsphere-ui-backend/pom.xml
@@ -53,12 +53,12 @@
         </dependency>
         <dependency>
             <groupId>org.apache.shardingsphere</groupId>
-            <artifactId>shardingsphere-orchestration-center-zookeeper-curator</artifactId>
+            <artifactId>shardingsphere-orchestration-repository-zookeeper-curator</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
             <groupId>org.apache.shardingsphere</groupId>
-            <artifactId>shardingsphere-orchestration-center-etcd</artifactId>
+            <artifactId>shardingsphere-orchestration-repository-etcd</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
