+++ shardingsphere-ui-backend/src/main/java/org/apache/shardingsphere/ui/servcie/RegistryCenterService.java
@@ -17,9 +17,8 @@
 
 package org.apache.shardingsphere.ui.servcie;
 
-import org.apache.shardingsphere.orchestration.center.RegistryCenterRepository;
-import org.apache.shardingsphere.orchestration.core.configcenter.ConfigCenterNode;
-import org.apache.shardingsphere.orchestration.core.registrycenter.RegistryCenterNode;
+import org.apache.shardingsphere.orchestration.core.registry.RegistryCenterNode;
+import org.apache.shardingsphere.orchestration.repository.api.RegistryCenterRepository;
 
 /**
  * Registry center service.
